Xqci: Fix unintended overflows
Fixes a bug in

  Xreg width = width_minus1 + 1,

where width_minus1 is a Bits<N> type with N smaller than 32. Since the
C literal "1" is 1 bit in size it will be extended to N bits and the
N-bit addition might overflow. The range of width is then [1,31]
instead of [1,32] as intended. Fixed by replacing add with a widening
add.

Signed-off-by: Anton Johansson <anjo@rev.ng>